Update Main.py
Added a While loop Statement for when asking for lucky number. So now the user can only enter numbers 1 through 10 and if anything else is imputed, it will give an error and ask them to enter the number again. It will keep looping until the correct range of numbers is entered.